+++ src/main/java/subway/api/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.api.dto;
 
 public class StationRequest {
     private String name;
