+++ src/main/java/subway/api/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.api.dto;
 
 public class StationResponse {
     private Long id;
