+++ src/main/java/subway/domain/station/repository/StationRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.domain.station.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.station.entity.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
