+++ src/main/java/subway/domain/station/service/StationService.java
@@ -1,7 +1,11 @@
-package subway;
+package subway.domain.station.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.api.dto.StationRequest;
+import subway.api.dto.StationResponse;
+import subway.domain.station.entity.Station;
+import subway.domain.station.repository.StationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
